continuous patrol robust in sim
Fails in lab. Scanner generates multiple inf-s and this disrupts the
clear span definition loop and subsequently the sorting. As a result the robot
as it as it picks a nearby obstacle to head toward. Also, the clear
spans are all very narrow and the first discontinuity is RISE. It is
very possible that this case is erroneous and it actually flips the
obstacles and clear spans. To test!