classless direction service test doesn't work
It requies more knowledge of executors and multi-threaded ROS 2. Tried
to add a scan subscriber node to the main program but then that also has
to be spun up and so spin_until_future_complete cannot be called. If the
scan sub can be spun up in a separate thread forked from main, then it
will be straightforward to use the response callback lambda when calling
async_send_request.